lead_flow_backend/requirements.txt
asgiref==3.11.1
coverage==7.13.5
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2025.3
